[feat] 이미지 캐시 구현
Extensions 코드 위치 이동
DetailVC 이미지 아이콘 Constraint 수정
